Revision and clean up to the deleteAll function

+++ listAndDelete.py
@@ -1,28 +1,17 @@
 import boto3
+import botocore
 
-#Retrieves the list of existing buckets
-s3 = boto3.client('s3')
-s3Resource = boto3.resource('s3')
+client = boto3.client('s3')
+s3 = boto3.resource('s3')
 
+buckets = client.list_buckets()
 
-response = s3.list_buckets()
-#print(type(response))
-
-#List to collect all bucket names
-lst = list()
-
-#Outputs the bucket names
-print("Existing Bucket")
-for bucket in response['Buckets']:
-    #print(f' {bucket["Name"]}')
-    lst.append(bucket["Name"])
-
-#print(lst)
-
-def deletelistofbuckets(lst):
-    for bucket in lst:
-        s3_bucket = s3Resource.Bucket(bucket)
-        s3_bucket.objects.all().delete()
-        print(f'Deleting all objects in {bucket}')
-
-deletelistofbuckets(lst)
+for bucket in buckets['Buckets']:
+    if(bucket['Name']=="awspersonalbucketpeternguyen"):
+        continue
+    s3_bucket = s3.Bucket(bucket['Name'])
+    print(f"buckets currently getting deleted: {bucket}")
+    s3_bucket.objects.all().delete()
+    s3_bucket.objects_versions.delete()
+    s3_bucket.delete()
+
